change default prompt from sorin to damoekri
Numereous issues have been created due to zombie processes, failing
async, and other issues that are brought about by the sorin prompt's use
of async (as well as the pure prompt). These issues were largely fixed
in zsh 5.1, but many people are using an older version.

To proactively reduce these issues, change the default prompt from sorin
to damoekri, which is extremely similar in apperance and behavior. The
main differentiating factor is the lack of async git support, which is
the exact thing that causes many of these issues.

Further into the future, when many have updated to a newer version of
zsh (like when Apple bundles 5.1 in their next release), it may be wise
to either revert this change, or add async support to other themes.