update submodules
Updates mode to the pure prompt have been made in an attempt to fix some
async issues.

sindresorhus/pure@9de635e fixes a git pull issue.
sindresorhus/pure@dc30cc6 is an update to the async code that both the
pure and sorin prompt rely on. This does not seem to have caused any
regressions. This also may 'fix' (workaround) some issues with async
support for zsh < 5.1.

More testing of this will be needed before the default prompt theme is
reverted to a prompt with async support, specifically testing of zsh <
5.1.