Add 'editor-ignores' to completion
This adds an option to the completion module that allows the editor tab
completion to ignore common binary/media file types. This option must
be explicitly enabled, as some people may use their text editors as hex
editors. Though, this is likely not many people, so it is enabled by
default for a new install in runcoms/zpreztorc.

This closes #10.